[PyTorch] Optimize TupleType::annotation_str_impl
stringstream is expensive to create, we used stringstream instead of ostringstream, and we can easily specialize the empty tuple. Also, anybody compiling with C++20 support can move out of the stringstream and it shouldn't hurt people without C++20 support to do so. I would consider specializing the 1-element case as well but I don't have evidence that that's necessary right now.

Differential Revision: [D43882402](https://our.internmc.facebook.com/intern/diff/D43882402/)

ghstack-source-id: 182186345
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96215