Rewrite fallthrough to more closely match how C++ works
Fallthrough is modeled as a mask which we use to remove keys from the
compute dispatch key set for eligibility.

It's possible this addresses https://github.com/pytorch/pytorch/issues/89037
in a better way than https://github.com/pytorch/pytorch/pull/95891 but I
cannot easily tell as the original repro no longer works and the new PR
does not have a test.

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: 89ca0317b891bb0cfdb111e722850579bdc22656
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96304