Add merge_rule for "functorch" module
This PR enables our non-meta contributors to be able to approve
"functorch" PRs without intervention from meta contributors.

A PR is deemed a "functorch" PR if it matches one of the patterns in
merge_rules.yaml. These patterns are definitely not exhaustive
(we modify core pytorch pieces quite often), but should be a good starting
point.

ghstack-source-id: d7e4ede0af610341a8f13ed78e5b302f33f4444a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96657