[memory profiling] combined stack API
This refactors the stack trace facility specific to memory profiling
in python+cuda to make a generic facility to generate combined stack
traces.

The generate facility (combined_traceback.h) does not require
python to be around to work, but will return python stacks if it is
present.

This facility is then used to add support for stack trace gathering in memory profiling that
happens directly from C++.

It is also used to expose a python API for gathering and symbolizing
combineds stacks.

ghstack-source-id: f9cf6a1b23155fa1b4d0f58b769202893db9cd9d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/95541

[memory profiling] python API for gather combined stacks

Refactors the code that generates combined python/c++/script/inductor
stacks into a separate file, and adds a direct interface to gather
a traceback and symbolize it.

ghstack-source-id: f9cf6a1b23155fa1b4d0f58b769202893db9cd9d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96281