[allocator] Generalize recording to a pool
Previously the allocator would query whether a stream was recording a graph,
and look up the pool associated with a graph. This change has the allocator
directly associate a stream with a mempool, decoupling "record this stream to a pool"
from the action of "record all actions to a cuda graph".

ghstack-source-id: c3cf84438b1a21a9626510f231bdec5b1d678201
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96542