[memory profiling] show pre-existing memory in trace_plot
Previously we only plotted memory if it was allocated or freed while
trace recording was active. This change also adds any pre-existing blocks
to the visualization. This helps because it is common to enable trace recording
later and then not realize that there is a lot of allocated memory in
the trace eventhough a lot was allocated beforehad.

ghstack-source-id: d2747c62f63903897452112b63725adc9dbd4775
Pull Request resolved: https://github.com/pytorch/pytorch/pull/97590