[PyTorch] Optimize TupleType::annotation_str_impl for small tuples
Pull Request resolved: https://github.com/pytorch/pytorch/pull/97910

In general, we can't profitably gather an array of all the elements' annotation strings so that we can reserve the final string because we'd have to heap-allocate that array. If we do it as a fast path for small tuples (which Tuple itself sets precedent for!), we can stack-allocate the array of annotation strings and make it profitable.
ghstack-source-id: 184566238

Differential Revision: [D44519675](https://our.internmc.facebook.com/intern/diff/D44519675/)