Clarify the saving of intermediates in the "extending torch.func" docs
Fixes https://github.com/pytorch/pytorch/issues/97260

We got some feedback that the page reads like "in order to save an input
for backward, you must return it as an output of the
autograd.Function.forward".

Doing so actually raises an error (on master and as of 2.1), but results
in an ambiguous situation on 2.0.0. To avoid more users running into
this, we clarify the documentation so it doesn't read like the above
and clearly mentions that you can save things from the inputs or
outputs.

ghstack-source-id: 50153883c5391684c748e9e7d74fb8f138e3a97a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98020