[inductor] Add an AOT mode for the Triton backend
Summary: This PR implements a two-pass wrapper codegen for the Triton
backend to achieve ahead-of-time compilation. In the first pass, the
regular python wrapper code will be generated, and then the generated
code will be executed to perform Triton compilation and autotuning.
After that, the second pass wrapper codegen will generate C++ wrapper
with proper CUDA API to load and launch Triton-generated CUDA kernels.

Like the AOT mode for the cpp backend, the next step would be to provide
a more complete API for AOT.

ghstack-source-id: ac1d54411b77c36f179065cdb4d9f313a7fd48cd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/97152