Improve the performance of softmax in the mid range
We choose to go for a large block size. In any of the casees, the number of blocks per SM is 1,
so this increases the parallelism. Also, the PTX code seems to be more efficient
(the memory throughput goes from 69% to 77%). To go with this, we also increment the max RBLOCK.

ghstack-source-id: 85a96b3ae7adf5a02cd1c98817c083b00ec10424
Pull Request resolved: https://github.com/pytorch/pytorch/pull/90994