Cleanup collectives and inductor backend.
Rename ForceInplace to InPlaceHint and use it with all collectives.
Unify codegen of AllReduceCoallesced with others.

Fix the wait behavior of all_reduce_coallesced so that you only need to
wait on a single tensor even when code-gen'ing. This requires ref-counting
the wait cleanup to ensure it won't trigger a wait too soon.

ghstack-source-id: ab1a11e9bf000b7f0560db6e65e9e112c86ace41
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98554