[torch.library] Add ability to create library fragments
In C++ we have TORCH_LIBRARY_FRAGMENT. This PR adds the same
functionality to the Python torch.library API.

The motivation for this is: for the simple custom op API, we don't want
users to need to deal with Library objects. One way to hide this from
users is to create library fragments.

Test Plan:
- tests that you can create multiple fragments and def+impl operators on each.

ghstack-source-id: 96d4d6f9394c0fb6c176327fc438d43b6b422b33
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98439