Make Adam optimizer (almost) work
Currently, use faketensor execution for the first optimizer step,
to create correct-shaped fake tensors for optimizer states, and
avoid taking the init path on the second try.

Then we hackily convert the optimizer states to zero tensors.  This is the
worst part of the current impl, as it fails to trace the real optimizer
init and could be incorrect.

Finally, we can wrap the real optimizer state tensors with fake tensors
to perform a real train step trace.

ghstack-source-id: 5fdee64f49824e876ab5b68f53b3c78759c05077
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98647