Respect sharded dimensions when aten expaned/view consumes SymInt values
Currently, aten.expand always expands to the global dimension. Then, it
introduces additional slice and clone ops before running compute on
the expanded tensor with a local tensor.

In this commit, if we detect the op consumes a SymInt size, it respects
both local size and the dimension placements from where the SymInt was
extracted.

ghstack-source-id: f1a6eadb0177ec2f020a290797ead5b95df09c22
Pull Request resolved: https://github.com/pytorch/pytorch/pull/99058