Remove IS_CI check from inductor lowering
The IS_CI check was designed to only put a hard error in CI and not bother users.

The warning that predates the hard error was bothering users becuase
 1- it was not actionable for the users
 2- it was introduced by changes made by other devs adding decomps

Making it a hard error closes the loop for the devs adding decomps, forcing action
at that time.  In this regime, users shouldn't see the warning unless they also
added a decomp.

It's still possible users would hit this warning on an implicit fallback if we did not
test that op in our CI at all.  This seems relatively low risk.

ghstack-source-id: b7ace0d4c80591a2b81ccce44062a31dccc0b5eb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/99995