[CustomOp] Fix lifetime semantics
This PR makes a CustomOp live forever. The motivation for it living
forever is that:
1. It doesn't matter to a user if it lives forever or not
2. it is a higher-level abstraction over OpOverload, and OpOverload
assumes that OpOverload lives forever.

The only place where it matters that CustomOp lives forever is testing:
I don't want to generate random names for my CustomOp objects. To
resolve the testing problem, This PR adds a CustomOp._destroy() that
clears all the C++ state, including the OpOverloadPacket, that is
associated with the CustomOp object.

Test Plan:
- existing tests

ghstack-source-id: d17ca8d3c35d5a3c022e946f37d7e9bea27a15bd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/100114