Refactor OptimizeIndexing to make it more extensible
This PR decouples the logic necessary to compute bounds on variables
from the logic that uses this info to perform the strenght analysis on
int64 variables. While doing so, it tries to minimize the number of
attributes of the class in favour of local variables.

This class is now accessible from any `LoopBody` object.

ghstack-source-id: 6bea057a373c2a8f3f616c10b0685c9b52fa3cfa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/100549